Updated and fixed WordAnalysisTest class so that it would work with the new abstract class for storing the users input. This involved creating a new constructor in the WordAnalysis class so that the test text, being inputted as a test parameter, could be passed into the abstract class from within the WordAnalysis class.
Also removed the need to test the average sentences per paragraph calculation as this has been removed from the application.

This test class also needed to initialise the JavaFx framework (since moving from a TUI to GUI). I have added in code to do this. This code will also need to be added in to every other test script which refers to the GUI (i.e. the static variables and where classes extend the abstract class).

Finally, the name of the folder where test data is stored has been renamed so I have amended the references to this to reflect the correct file path.